extract colored hats logic into separate file
