+++ STROOP/STROOP.csproj
@@ -425,6 +425,9 @@
     </Compile>
     <Compile Include="Tabs\BruteforceTab\BF_Utilities\BF_ObjectUtilities.cs" />
     <Compile Include="Tabs\BruteforceTab\ValueGetters.cs" />
+    <Compile Include="Tabs\GhostTab\ColoredHats.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
     <Compile Include="Tabs\GhostTab\FileWatchers.cs">
       <SubType>UserControl</SubType>
     </Compile>
